refactor(calendar): extract CalendarHeader from CalendarNavbar
Moves getDateRange and its Day/Week/Month formatting logic out of
CalendarNavbar into a dedicated CalendarHeader component.

formatMeetingDateLine now takes an isHeader flag (default false) and
only shows the "· Today" badge when isHeader is true, so it no longer
appears on every caller (e.g. the ViewMeeting popup) -- only the
calendar header opts in.
